more messages

+++ slurminator.sh
@@ -33,17 +33,25 @@ function get_job_name {
 function kill_multiple_jobs {
 	TJOBS=$(get_squeue_from_format_string "'%A' '%j (%t, %M)' OFF")
 	chosenjobs=$(eval "whiptail --title 'Which jobs to kill?' --checklist 'Which jobs to choose?' $WIDTHHEIGHT $TJOBS" 3>&1 1>&2 2>&3)
-	if (whiptail --title "Really kill multiple jobs ($chosenjobs)?" --yesno "Are you sure you want to kill multiple jobs ($chosenjobs)?" 8 78); then
-		echo "scancel $chosenjobs"
-		eval "scancel $chosenjobs"
+	if [[ $chosenjob -e "" ]]; then
+		green_text "No jobs chosen to kill"
+	else
+		if (whiptail --title "Really kill multiple jobs ($chosenjobs)?" --yesno "Are you sure you want to kill multiple jobs ($chosenjobs)?" 8 78); then
+			echo "scancel $chosenjobs"
+			eval "scancel $chosenjobs"
+		fi
 	fi
 }
 
 function tail_multiple_jobs {
 	TJOBS=$(get_squeue_from_format_string "'%A' '%j (%t, %M)' OFF")
-	test=$(eval "whiptail --title 'Which jobs to tail?' --checklist 'Which jobs to choose?' $WIDTHHEIGHT $TJOBS" 3>&1 1>&2 2>&3)
+	chosenjobs=$(eval "whiptail --title 'Which jobs to tail?' --checklist 'Which jobs to choose?' $WIDTHHEIGHT $TJOBS" 3>&1 1>&2 2>&3)
 	whiptail --title "Tail for multiple jobs with screen" --msgbox "To exit, press <CTRL> <a>, then <\\>" 8 78 3>&1 1>&2 2>&3
-	eval "multiple_slurm_tails $test"
+	if [[ $chosenjobs -e "" ]]; then
+		green_text "No jobs chosen to tail"
+	else
+		eval "multiple_slurm_tails $chosenjobs"
+	fi
 }
 
 function single_job_tasks {
